Add dict encoder/decoder
Resolves #921
