+++ xsdata/formats/dataclass/parsers/__init__.py
@@ -1,5 +1,12 @@
+from xsdata.formats.dataclass.parsers.dict import DictDecoder
 from xsdata.formats.dataclass.parsers.json import JsonParser
 from xsdata.formats.dataclass.parsers.tree import TreeParser
 from xsdata.formats.dataclass.parsers.xml import UserXmlParser, XmlParser
 
-__all__ = ["JsonParser", "XmlParser", "UserXmlParser", "TreeParser"]
+__all__ = [
+    "DictDecoder",
+    "JsonParser",
+    "XmlParser",
+    "UserXmlParser",
+    "TreeParser",
+]
