+++ xsdata/formats/dataclass/serializers/__init__.py
@@ -1,8 +1,10 @@
 from xsdata.formats.dataclass.serializers.code import PycodeSerializer
+from xsdata.formats.dataclass.serializers.dict import DictEncoder
 from xsdata.formats.dataclass.serializers.json import DictFactory, JsonSerializer
 from xsdata.formats.dataclass.serializers.xml import XmlSerializer
 
 __all__ = [
+    "DictEncoder",
     "DictFactory",
     "JsonSerializer",
     "XmlSerializer",
